add parallel support

+++ code/projects/TC_Tracker/core/include/multiArray.h
@@ -68,7 +68,7 @@ public:
         auto maxElemIndex = std::distance(data.get(), maxElemIter);
         return {{maxElemIndex/(_rows*_columns), maxElemIndex%(_rows*_columns)/_columns, maxElemIndex%_columns}, *maxElemIter};
     }
-    std::pair<std::pair<int, int>, float> max(int index) const {
+    std::pair<std::pair<int, int>, float> max(size_t index) const {
         auto maxElemIter = std::max_element(data.get()+index*_rows*_columns, data.get()+(index+1)*_rows*_columns);
         auto maxElemIndex = std::distance(data.get()+index*_rows*_columns, maxElemIter);
         return {{maxElemIndex/_columns, maxElemIndex%_columns}, *maxElemIter};
